pyproject.toml
[project]
name = "jakcy-rag-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bcrypt==4.1.3",
    "chromadb>=1.3.0",
    "fastapi>=0.120.3",
    "langchain==1.0.3",
    "langchain-chroma>=1.0.0",
    "langchain-core>=1.0.2",
    "langchain-huggingface>=1.0.0",
    "langchain-openai>=1.0.1",
    "langchain-text-splitters>=1.0.0",
    "numpy>=2.3.4",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "passlib[bcrypt]>=1.7.4",
    "pymupdf>=1.26.5",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sentence-transformers>=5.1.2",
]
